Fix chain switching
