+++ Sources/Web3Modal/Core/Web3ModalClient.swift
@@ -215,10 +215,13 @@ public class Web3ModalClient {
                     switch result {
                     case let .success(payload):
                         
-                        if let actionResult: String = try? payload.content.first?.get().rawValue {
-                            response = .init(result: .response(AnyCodable(any: actionResult)))
-                        } else {
-                            response = .init(result: .error(.init(code: -1, message: "Invalid response")))
+                        switch payload.content.first {
+                        case let .success(JSONString):
+                            response = .init(result: .response(AnyCodable(JSONString)))
+                        case let .failure(error):
+                            response = .init(result: .error(.init(code: error.code, message: error.message)))
+                        case .none:
+                            response = .init(result: .error(.init(code: -1, message: "Empty response")))
                         }
                     case let .failure(error):
                         Web3Modal.config.onError(error)
@@ -332,85 +335,4 @@ public class Web3ModalClient {
             return false
         }
     }
-}
-
-// MARK: - Mapping
-
-extension W3MJSONRPC {
-    func toCbAction() -> Web3JSONRPC? {
-        switch self {
-        case let .personal_sign(address, message):
-            return .personal_sign(
-                address: address,
-                message: message
-            )
-        case .eth_requestAccounts:
-            return .eth_requestAccounts
-        case let .eth_signTransaction(from, to, value, data, nonce, _, gasPrice, maxFeePerGas, maxPriorityFeePerGas, gasLimit, chainId):
-            return .eth_signTransaction(
-                fromAddress: from,
-                toAddress: to,
-                weiValue: value,
-                data: data,
-                nonce: nonce,
-                gasPriceInWei: gasPrice,
-                maxFeePerGas: maxFeePerGas,
-                maxPriorityFeePerGas: maxPriorityFeePerGas,
-                gasLimit: gasLimit,
-                chainId: chainId
-            )
-        case let .eth_sendTransaction(from, to, value, data, nonce, _, gasPrice, maxFeePerGas, maxPriorityFeePerGas, gasLimit, chainId):
-            return .eth_sendTransaction(
-                fromAddress: from,
-                toAddress: to,
-                weiValue: value,
-                data: data,
-                nonce: nonce,
-                gasPriceInWei: gasPrice,
-                maxFeePerGas: maxFeePerGas,
-                maxPriorityFeePerGas: maxPriorityFeePerGas,
-                gasLimit: gasLimit,
-                chainId: chainId
-            )
-        case let .wallet_switchEthereumChain(chainId):
-            return .wallet_switchEthereumChain(chainId: chainId)
-        case let .wallet_addEthereumChain(chainId, blockExplorerUrls, chainName, iconUrls, nativeCurrency, rpcUrls):
-            return .wallet_addEthereumChain(
-                chainId: chainId,
-                blockExplorerUrls: blockExplorerUrls,
-                chainName: chainName,
-                iconUrls: iconUrls,
-                nativeCurrency: nativeCurrency != nil ? .init(
-                    name: nativeCurrency!.name,
-                    symbol: nativeCurrency!.symbol,
-                    decimals: nativeCurrency!.decimals
-                ) : nil,
-                rpcUrls: rpcUrls
-            )
-        case let .wallet_watchAsset(type, options):
-            return .wallet_watchAsset(
-                type: type,
-                options: .init(
-                    address: options.address,
-                    symbol: options.symbol,
-                    decimals: options.decimals,
-                    image: options.image
-                )
-            )
-        }
-    }
-}
-
-public struct W3MResponse: Codable {
-    init(id: RPCID? = nil, topic: String? = nil, chainId: String? = nil, result: RPCResult) {
-        self.id = id
-        self.topic = topic
-        self.chainId = chainId
-        self.result = result
-    }
-    
-    public let id: RPCID?
-    public let topic: String?
-    public let chainId: String?
-    public let result: RPCResult
 }
